Make that zipped CSV export reusable
I'm going to need the exact same thing for another export.

Relates to #595
